requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
tqdm
pyyaml
pretty_midi
mido
streamlit
Pillow
opencv-python
transformers
torchvision
soundfile
scipy
regex
keyboard
fluidsynth
